training/requirements.txt
absl-py==0.9.0
astor==0.8.1
boto3==1.10.45
botocore==1.13.45
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
feedparser==5.2.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
jmespath==0.9.4
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
pkg-resources==0.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
sacremoses==0.0.35
sentencepiece==0.1.85
six==1.13.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
tqdm==4.41.1
transformers==2.3.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
